Add support for mysql bulk update

+++ src/lib/doctrine-dbal-bulk/tests/Flow/Doctrine/Bulk/Tests/Context/DatabaseContext.php
@@ -39,6 +39,10 @@ final class DatabaseContext
     public function dropAllTables() : void
     {
         foreach ($this->connection->getSchemaManager()->listTables() as $table) {
+            if (str_contains($table->getName(), 'innodb')) {
+                continue;
+            }
+
             $this->connection->getSchemaManager()->dropTable($table->getName());
         }
     }
